Refactor ConstructionAssignmentSource to use enum type
Previously, `DataSurfaces::SurfaceData::ConstructionAssignmentSource` was stored as an `int` mirroring `ConstructionAssignments::SearchDistanceType`. This commit changes it to directly use the enum, improving type safety.

Display strings for the enum values are now centralized in `ConstructionAssignmentSet.hh`, simplifying reporting logic and reducing redundancy.

+++ src/EnergyPlus/SurfaceGeometry.cc
@@ -2491,7 +2491,7 @@ namespace SurfaceGeometry {
             for (int SurfNum = 1; SurfNum <= state.dataSurface->TotSurfaces; ++SurfNum) {
                 auto &surf = state.dataSurface->Surface(SurfNum);
                 if (surf.HeatTransSurf && surf.Construction != 0) {
-                    surf.ConstructionAssignmentSource = static_cast<int>(ConstructionAssignments::SearchDistanceType::HardAssigned);
+                    surf.ConstructionAssignmentSource = ConstructionAssignments::SearchDistanceType::HardAssigned;
                 }
             }
 
@@ -2540,7 +2540,7 @@ namespace SurfaceGeometry {
                 auto &surf = state.dataSurface->Surface(SurfNum);
                 surf.Construction = resolvedConstrNum[SurfNum];
                 surf.ConstructionStoredInputValue = resolvedConstrNum[SurfNum];
-                surf.ConstructionAssignmentSource = static_cast<int>(resolvedSource[SurfNum]);
+                surf.ConstructionAssignmentSource = resolvedSource[SurfNum];
                 state.dataConstruction->Construct(surf.Construction).IsUsed = true;
             }
         }
